refactor: extract cli/
chore: upgrade deno version
