+++ src/incus-app-container-files/cli/create-cli.ts
@@ -1,20 +1,20 @@
 import { isString } from "https://deno.land/x/fns@1.1.0/string/is-string.ts";
-import { AbsolutePath, isAbsolutePath } from "./absolute-path.ts";
-import { Config } from "./config.ts";
+import { AbsolutePath, isAbsolutePath } from "../absolute-path.ts";
+import { Config } from "../config.ts";
 import {
   CreateAppContainerOptions,
   resolveCreateAppContainerOptions,
-} from "./create-app-container-options.ts";
-import { createAppContainer } from "./create-app-container.ts";
-import { breadc, run, s } from "./deps.ts";
+} from "../create-app-container-options.ts";
+import { createAppContainer } from "../create-app-container.ts";
+import { breadc, run, s } from "../deps.ts";
 import {
   INCUS_CONTAINER_STATUS_CODES,
   untilStatusCode,
-} from "./incus-container-status.ts";
-import { isOutputFormat, OUTPUT_FORMATS } from "./output-format.ts";
-import { isSize } from "./size.ts";
-import { isSshKey } from "./ssh-key.ts";
-import { enforceType, optional } from "./type-guard.ts";
+} from "../incus-container-status.ts";
+import { isOutputFormat, OUTPUT_FORMATS } from "../output-format.ts";
+import { isSize } from "../size.ts";
+import { isSshKey } from "../ssh-key.ts";
+import { enforceType, optional } from "../type-guard.ts";
 
 export const COMMAND_NAMES = ["create", "delete", "list"] as const;
 export type CommandName = typeof COMMAND_NAMES[number];
