+++ src/incus-app-container-files/cli/mod.ts
@@ -6,9 +6,9 @@
  * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-import { Config, getConfig } from "./config.ts";
+import { Config, getConfig } from "../config.ts";
 import { createCli } from "./create-cli.ts";
-import { CommandFailureError, ParseError } from "./deps.ts";
+import { CommandFailureError, ParseError } from "../deps.ts";
 
 const config: Config = await getConfig();
 const cli = await createCli(config);
